+++ c/randoms.c
@@ -6,7 +6,7 @@ double random_double() {
     return (rand() + 1.0) / (RAND_MAX + 1.0);
 }
 
-double random_exponential(int rho) {
+double random_exponential(double rho) {
     return -log(random_double) / rho;
 }
