+++ c/randoms.h
@@ -1,7 +1,7 @@
 #ifndef RANDOMS
 #define RANDOMS
 
-double random_exponential(int rho);
+double random_exponential(double rho);
 double poisson_quantile(double epsilon);
 
 #endif
